+++ pyogrio/_compat.py
@@ -18,10 +18,17 @@ try:
 except ImportError:
     geopandas = None
 
+try:
+    import pandas
+except ImportError:
+    pandas = None
+
 
 HAS_ARROW_API = __gdal_version__ >= (3, 6, 0) and pyarrow is not None
 
 HAS_GEOPANDAS = geopandas is not None
+
+PANDAS_GE_20 = pandas is not None and Version(pandas.__version__) >= Version("2.0.0")
 
 HAS_GDAL_GEOS = __gdal_geos_version__ is not None